3/test_three.py
import three


def test1():
    input_data = (
        ['R75', 'D30', 'R83', 'U83', 'L12', 'D49', 'R71', 'U7', 'L72'],
        ['U62', 'R66', 'U55', 'R34', 'D71', 'R55', 'D58', 'R83'],
    )
    assert three.get_distance_to_closest_cross(input_data) == 159


def test2():
    input_data = (
        ['R98', 'U47', 'R26', 'D63', 'R33', 'U87', 'L62', 'D20', 'R33', 'U53', 'R51'],
        ['U98', 'R91', 'D20', 'R16', 'D67', 'R40', 'U7', 'R15', 'U6', 'R7'],
    )
    assert three.get_distance_to_closest_cross(input_data) == 135


def test3():
    input_data = (
        ['R75', 'D30', 'R83', 'U83', 'L12', 'D49', 'R71', 'U7', 'L72'],
        ['U62', 'R66', 'U55', 'R34', 'D71', 'R55', 'D58', 'R83'],
    )
    assert three.get_wire_distance_to_closest_cross(input_data) == 610


def test4():
    input_data = (
        ['R98', 'U47', 'R26', 'D63', 'R33', 'U87', 'L62', 'D20', 'R33', 'U53', 'R51'],
        ['U98', 'R91', 'D20', 'R16', 'D67', 'R40', 'U7', 'R15', 'U6', 'R7'],
    )
    assert three.get_wire_distance_to_closest_cross(input_data) == 410
